kiara/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
cryptography==2.4.1
Django==2.1.3
django-filter==2.0.0
django-rest-knox==3.4.0
djangorestframework==3.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pyOpenSSL==18.0.0
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
wrapt==1.10.11
gunicorn==19.7.1

